moved dump statement

+++ src/corewar/src/op_bit.c
@@ -11,7 +11,7 @@ void	xor(uint8_t core[MEM_SIZE], t_carriage **carriage, t_info *info)
 	if (print_command(info) == TRUE && info)
 		v_flag4_three_arg(carriage, "xor", ARG3);
 	sum = (*carriage)->args_found[ARG1] ^ (*carriage)->args_found[ARG2];
-	(*carriage)->registry[(*carriage)->args_found[ARG3] - 1] = sum;
+	(*carriage)->registry[(*carriage)->args_found[ARG3] - 1] = (int32_t)sum;
 	update_carry(sum, carriage);
 	if (info->flag[V_FLAG] == 25 && found == FALSE)
 		found = v_flag5(carriage);
@@ -27,7 +27,7 @@ void	or(uint8_t core[MEM_SIZE], t_carriage **carriage, t_info *info)
 	if (print_command(info) == TRUE && info)
 		v_flag4_three_arg(carriage, "or", ARG3);
 	sum = (*carriage)->args_found[ARG1] | (*carriage)->args_found[ARG2];
-	(*carriage)->registry[(*carriage)->args_found[ARG3] - 1] = sum;
+	(*carriage)->registry[(*carriage)->args_found[ARG3] - 1] = (int32_t)sum;
 	update_carry(sum, carriage);
 	if (info->flag[V_FLAG] == 25 && found == FALSE)
 		found = v_flag5(carriage);
@@ -43,7 +43,7 @@ void	and(uint8_t core[MEM_SIZE], t_carriage **carriage, t_info *info)
 	if (print_command(info) == TRUE && info)
 		v_flag4_three_arg(carriage, "and", ARG3);
 	sum = (*carriage)->args_found[ARG1] & (*carriage)->args_found[ARG2];
-	(*carriage)->registry[(*carriage)->args_found[ARG3] - 1] = sum;
+	(*carriage)->registry[(*carriage)->args_found[ARG3] - 1] = (int32_t)sum;
 	update_carry(sum, carriage);
 	if (info->flag[V_FLAG] == 25 && found == FALSE)
 		found = v_flag5(carriage);
