+++ src/corewar/src/op_helper.c
@@ -56,8 +56,8 @@ void	update_arg_values(uint8_t core[MEM_SIZE], t_carriage **carriage, int64_t *a
 	else if ((*carriage)->arg_types[count] == I)
 	{
 		if (*arg < 0)
-			*arg = read_bytes(0, (*carriage)->pos - ((*arg * -1) % IDX_MOD), core, SIZE);
+			*arg = (int32_t)read_bytes(0, (*carriage)->pos - ((*arg * -1) % IDX_MOD), core, SIZE);
 		else
-			*arg = read_bytes(0, (*carriage)->pos + (*arg % IDX_MOD), core, SIZE);
+			*arg = (int32_t)read_bytes(0, (*carriage)->pos + (*arg % IDX_MOD), core, SIZE);
 	}
 }
